requirements.txt
fastapi==0.85.0
pydantic==1.10.13
uvicorn==0.18.2
motor==2.5.1
pandas==1.4.2
openpyxl==3.0.9
python-dotenv==0.19.2
